+++ src/services/notificationService.ts
@@ -1,0 +1,5 @@
+import api from "./api";
+
+export const markNotificationAsRead = async (notificationId: number) => {
+  return api.patch(`/api/v1/notifications/${notificationId}/read`);
+};
