+++ src/types/member.d.ts
@@ -1,0 +1,23 @@
+type SignInRequest = {
+  email: string;
+  password: string;
+};
+
+type SignUpRequest = {
+  email: string;
+  password: string;
+  birthday: string;
+  nickname: string;
+  img: string;
+  phoneNumber: string;
+  role: string;
+  hostBizNumber: string;
+  hostBizCeoName: string;
+  hostBizName: string;
+  hostBizAddress: string;
+  hostBizEcommerceRegistrationNumber: string;
+  hostBizBankName: string;
+  hostBizDepositor: string;
+  hostBizBankNumber: string;
+  hostContractCharge: number;
+};
