+++ src/types/notification.d.ts
@@ -1,0 +1,7 @@
+type NotificationResponse = {
+  id: number;
+  title: string;
+  content: string;
+  createdAt: string;
+  isRead: boolean;
+};
